postgres_to_es/requirements.txt
psycopg2-binary==2.9.7
backoff==2.2.1
pydantic==2.0.3
elasticsearch==8.8.2
redis==4.6.0
python-dotenv==1.0.0
